+++ tests/lib/rules/forbid-component-props.js
@@ -250,6 +250,78 @@ ruleTester.run('forbid-component-props', rule, {
         },
       ],
     },
+    {
+      code: `
+        const rootElement = (
+          <Root>
+            <SomeIcon className="size-lg" />
+            <AnotherIcon className="size-lg" />
+            <SomeSvg className="size-lg" />
+            <UICard className="size-lg" />
+            <UIButton className="size-lg" />
+          </Root>
+        );
+      `,
+      options: [
+        {
+          forbid: [
+            {
+              propName: 'className',
+              allowedForPatterns: ['*Icon', '*Svg', 'UI*'],
+            },
+          ],
+        },
+      ],
+    },
+    {
+      code: `
+        const rootElement = (
+          <Root>
+            <SomeIcon className="size-lg" />
+            <AnotherIcon className="size-lg" />
+            <SomeSvg className="size-lg" />
+            <UICard className="size-lg" />
+            <UIButton className="size-lg" />
+            <ButtonLegacy className="size-lg" />
+          </Root>
+        );
+      `,
+      options: [
+        {
+          forbid: [
+            {
+              propName: 'className',
+              allowedFor: ['ButtonLegacy'],
+              allowedForPatterns: ['*Icon', '*Svg', 'UI*'],
+            },
+          ],
+        },
+      ],
+    },
+    {
+      code: `
+        const rootElement = (
+          <Root>
+            <SomeIcon className="size-lg" />
+            <AnotherIcon className="size-lg" />
+            <SomeSvg className="size-lg" />
+            <UICard className="size-lg" />
+            <UIButton className="size-lg" />
+          </Root>
+        );
+      `,
+      options: [
+        {
+          forbid: [
+            {
+              propName: 'className',
+              disallowedFor: ['Modal'],
+              disallowedForPatterns: ['*Legacy', 'Shared*'],
+            },
+          ],
+        },
+      ],
+    },
   ]),
 
   invalid: parsers.all([
@@ -675,6 +747,127 @@ ruleTester.run('forbid-component-props', rule, {
           data: { prop: 'kebab-case-prop' },
           line: 6,
           column: 18,
+          type: 'JSXAttribute',
+        },
+      ],
+    },
+    {
+      code: `
+        const rootElement = () => (
+          <Root>
+            <SomeIcon className="size-lg" />
+            <SomeSvg className="size-lg" />
+          </Root>
+        );
+      `,
+      options: [
+        {
+          forbid: [
+            {
+              propName: 'className',
+              message: 'className available only for icons',
+              allowedForPatterns: ['*Icon'],
+            },
+          ],
+        },
+      ],
+      errors: [
+        {
+          message: 'className available only for icons',
+          line: 5,
+          column: 22,
+          type: 'JSXAttribute',
+        },
+      ],
+    },
+    {
+      code: `
+        const rootElement = () => (
+          <Root>
+            <UICard style={{backgroundColor: black}}/>
+            <SomeIcon className="size-lg" />
+            <SomeSvg className="size-lg" style={{fill: currentColor}} />
+          </Root>
+        );
+      `,
+      options: [
+        {
+          forbid: [
+            {
+              propName: 'className',
+              message: 'className available only for icons',
+              allowedForPatterns: ['*Icon'],
+            },
+            {
+              propName: 'style',
+              message: 'style available only for SVGs',
+              allowedForPatterns: ['*Svg'],
+            },
+          ],
+        },
+      ],
+      errors: [
+        {
+          message: 'style available only for SVGs',
+          line: 4,
+          column: 21,
+          type: 'JSXAttribute',
+        },
+        {
+          message: 'className available only for icons',
+          line: 6,
+          column: 22,
+          type: 'JSXAttribute',
+        },
+      ],
+    },
+    {
+      code: `
+        const rootElement = (
+          <Root>
+            <SomeIcon className="size-lg" />
+            <AnotherIcon className="size-lg" />
+            <SomeSvg className="size-lg" />
+            <UICard className="size-lg" />
+            <ButtonLegacy className="size-lg" />
+          </Root>
+        );
+      `,
+      options: [
+        {
+          forbid: [
+            {
+              propName: 'className',
+              disallowedFor: ['SomeSvg'],
+              disallowedForPatterns: ['UI*', '*Icon'],
+              message: 'Avoid using className for SomeSvg and components that match the `UI*` and `*Icon` patterns',
+            },
+          ],
+        },
+      ],
+      errors: [
+        {
+          message: 'Avoid using className for SomeSvg and components that match the `UI*` and `*Icon` patterns',
+          line: 4,
+          column: 23,
+          type: 'JSXAttribute',
+        },
+        {
+          message: 'Avoid using className for SomeSvg and components that match the `UI*` and `*Icon` patterns',
+          line: 5,
+          column: 26,
+          type: 'JSXAttribute',
+        },
+        {
+          message: 'Avoid using className for SomeSvg and components that match the `UI*` and `*Icon` patterns',
+          line: 6,
+          column: 22,
+          type: 'JSXAttribute',
+        },
+        {
+          message: 'Avoid using className for SomeSvg and components that match the `UI*` and `*Icon` patterns',
+          line: 7,
+          column: 21,
           type: 'JSXAttribute',
         },
       ],